requirements.txt
django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<25.0
openpyxl>=3.1,<4.0
